requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-multipart
openai
pydantic
python-dotenv
psycopg2-binary
gtts
